+++ install_defaults.php
@@ -54,7 +54,7 @@ $membershipConfigData = array(
     ),
     array(
         'name' => 'member_group',
-        'default_value' => '',
+        'default_value' => 13,
         'type' => 'select',
         'subgroup' => 0,
         'fieldset' => 0,
@@ -613,10 +613,9 @@ $membershipConfigData = array(
 /**
  * Initialize Membership plugin configuration.
  *
- * @param   integer $group_id   Group ID to use as the plugin's admin group
  * @return  boolean             true: success; false: an error occurred
  */
-function plugin_initconfig_membership($group_id = 0)
+function plugin_initconfig_membership()
 {
     global $membershipConfigData;